Education platform to analysis the Inventory
I have created a project on the Education platform to analyze the Inventory used in the last five years, how many of the students enrolled through application and also in online enrollment, how much software was used in the previous five years, how many of them using email, blogs in the five last year. By having this analysis, we can predict how much Inventory we need for the upcoming year. We can predict the student enrollment count too approximately.